+++ packages/extension/src/ui/features/shield/ShieldBaseOTPScreen.tsx
@@ -16,8 +16,8 @@ import {
   EmailVerificationStatus,
   getVerificationErrorMessage,
 } from "../../../shared/shield/backend/account"
-import { updateVerifiedEmail } from "../../../shared/shield/jwt"
 import { confirmEmail, requestEmail } from "../../../shared/shield/register"
+import { updateVerifiedEmail } from "../../../shared/shield/verifiedEmail"
 import { IS_DEV } from "../../../shared/utils/dev"
 import { coerceErrorToString } from "../../../shared/utils/error"
 import { ControlledPinInput } from "../../components/ControlledPinInput"
